+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 * Linking and including an external version of `fmt` is now supported. See TweakMe.h
 * Fixed compiler warnings when using clang's -Wdocumentation. Fixes [#12](https://github.com/odygrd/quill/issues/12).
 * Fixed a bug that wouldn't report a compile-time error for invalid format strings. Fixes [#13](https://github.com/odygrd/quill/issues/13).
+* Added process ID to Formatter. Fixes [#14](https://github.com/odygrd/quill/issues/14).
 
 ## v1.1.0
 * Daily file handler. The file handler rollover every 24 hours